Better structure on exports
